Separate settings to NanoGptSettings.cs Separate training to NanoGptTraining.cs Separate out main, to 'nano-gpt-cmd' folder Use kcg-xlib library for logging, io operations
